# openapi.yaml
openapi: 3.1.0
info:
  title: arca API
  version: v1
servers:
- url: https://arca-backend.dcism.org
  description: Production
- url: http://localhost:20255
  description: Local development
paths:
...
  /api/v1/votes:
    post:
      tags:
      - vote-controller
      operationId: vote
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/VoteRequest"
        required: true
      responses:
        "200":
          description: Vote saved
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/VoteResponse"
        "204":
          description: Existing vote removed
...
  /api/v1/votes/{postId}/upvotes:
    get:
      tags:
      - vote-controller
      operationId: getUpvoteCount
      parameters:
      - name: postId
        in: path
        required: true
        schema:
          type: integer
          format: int64
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                type: integer
                format: int32
  /api/v1/votes/{postId}/my-vote:
    get:
      tags:
      - vote-controller
      operationId: getMyVote
      parameters:
      - name: postId
        in: path
        required: true
        schema:
          type: integer
          format: int64
      responses:
        "200":
          description: Current vote
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/VoteResponse"
        "204":
          description: No current vote
  /api/v1/votes/{postId}/downvotes:
    get:
      tags:
      - vote-controller
      operationId: getDownvoteCount
      parameters:
      - name: postId
        in: path
        required: true
        schema:
          type: integer
          format: int64
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                type: integer
                format: int32
...
  /api/v1/votes/{postId}:
    delete:
      tags:
      - vote-controller
      operationId: removeVote
      parameters:
      - name: postId
        in: path
        required: true
        schema:
          type: integer
          format: int64
      responses:
        "204":
          description: Vote removed
...
components:
  schemas:
...
    VoteRequest:
      type: object
      properties:
        postId:
          type: integer
          format: int64
        voteType:
          type: string
    VoteResponse:
      type: object
      properties:
        id:
          type: integer
          format: int64
        postId:
          type: integer
          format: int64
        userId:
          type: integer
          format: int64
        voteType:
          type: string
...